[TS-409] Reduce code duplication
